+++ projects/angular-esri-components/src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of angular-esri-components
+ */
+
+export * from './lib/angular-esri-components.module';
+export * from './lib/core/esri-module-provider';