+++ projects/angular-esri-components/src/testing/esri-map.service.ts
@@ -1,0 +1,8 @@
+import { of } from 'rxjs';
+
+export class MockEsriMapSerivice {
+  isLoaded = of(true);
+  addWidget = jasmine.createSpy('addWidget');
+  loadMap = jasmine.createSpy('loadMap');
+  loadWebMap = jasmine.createSpy('loadWebMap');
+}