+++ projects/angular-esri-components/src/testing/index.ts
@@ -1,0 +1,3 @@
+export * from './esri-map.service';
+export * from './esri-module-provider';
+export * from './custom-widget.directive';