+++ src/lib/core/index.ts
@@ -1,2 +1,0 @@
-export * from './esri-map.service';
-export * from './esri-module-provider';